Saves Form Data, Time Passed and Heatmap in Server and Fixes Minor Bugs
Saves the Form data into the server file. Creates and saves also the passed time and a simple heatmap for each map in the server.
Fixes a bug where the amount of keys was not cleared after a level was finished. Fixed another bug in MyLizard level where a key had the wrong ID.
Fixes some UI bugs in the main screen. Made the font size bigger for the form and added a message so the user knows the level was cleared and the form must be answered. Translated everything to Portuguese.
Creates an ending screen so the user knows the game ended.